requirements.txt
streamlit
mysql-connector-python
pyresparser
pandas
python-docx
nltk
spacy
Pillow
